added vram initialization

+++ src/cpu.rs
@@ -21,15 +21,15 @@ impl Cpu {
     pub fn new() -> Self {
         Self {
             register: Register::new(),
-            stack: [0 as u16; STACK_SIZE],
+            stack: [0 as u16; STACK_SIZE as usize],
             stack_ptr: 0,
-            ram: [0 as u8; RAM_SIZE]
+            ram: [0 as u8; RAM_SIZE as usize]
             // Timer
             // Delay
 
         }
     }
-    /// Loads base fonts into RAM to interpreter section 0x000-0x1FF
+    /// Loads base fonts 0-10, A-F into interpreter section 0x000-0x1FF RAM
     pub fn init_font(&mut self){
         for i in 0..FONT_DEFAULTS.len(){
             self.ram[i] = FONT_DEFAULTS[i];
